+++ lib/containsDotFile.js
@@ -1,0 +1,23 @@
+/*!
+ * send
+ * Copyright(c) 2012 TJ Holowaychuk
+ * Copyright(c) 2014-2022 Douglas Christopher Wilson
+ * MIT Licensed
+ */
+'use strict'
+/**
+ * Determine if path parts contain a dotfile.
+ *
+ * @api private
+ */
+function containsDotFile (parts) {
+  for (let i = 0; i < parts.length; i++) {
+    const part = parts[i]
+    if (part.length > 1 && part[0] === '.') {
+      return true
+    }
+  }
+
+  return false
+}
+exports.containsDotFile = containsDotFile
