+++ lib/normalizeList.js
@@ -1,0 +1,23 @@
+'use strict'
+
+/**
+ * Normalize the index option into an array.
+ *
+ * @param {boolean|string|array} val
+ * @param {string} name
+ * @private
+ */
+
+function normalizeList (val, name) {
+  const list = [].concat(val || [])
+
+  for (let i = 0; i < list.length; i++) {
+    if (typeof list[i] !== 'string') {
+      throw new TypeError(name + ' must be array of strings or false')
+    }
+  }
+
+  return list
+}
+
+module.exports.normalizeList = normalizeList
