+++ lib/parseHttpDate.js
@@ -1,0 +1,18 @@
+'use strict'
+
+/**
+ * Parse an HTTP Date into a number.
+ *
+ * @param {string} date
+ * @private
+ */
+
+function parseHttpDate (date) {
+  const timestamp = date && Date.parse(date)
+
+  return typeof timestamp === 'number'
+    ? timestamp
+    : NaN
+}
+
+module.exports.parseHttpDate = parseHttpDate
